requirements.txt
pyspark
pandas
matplotlib
Pillow
torch
torchvision
tensorboard
jupyterlab
